code restyle + fix
